refactor: rename runlogic to runner and apply review fixes
- Rename runlogic/ to runner/ package for clarity
- Fix logging race condition: replace sync.Once with mutex nil-check
- Panic on missing viper in context (programming error)
- Propagate errors from run-scoped flag bindings
- Handle force flag error in config init
- Fix Error method formatting (args printed as slice)
- Add HTTP client timeout for ollama (5min)
- Log marshal errors in convertResponse instead of swallowing
- Fix checkRepeat ctx parameter order per Go convention
- Remove redundant normalizeProvider call in buildLLM
- Update Taskfile.yaml and agent prompts for runner rename
