+++ cmd/squad/config.go
@@ -30,7 +30,6 @@ import (
 
 	"github.com/cowdogmoo/squad/config"
 	"github.com/cowdogmoo/squad/logging"
-	"github.com/cowdogmoo/squad/runlogic"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	"gopkg.in/yaml.v3"
@@ -135,7 +134,10 @@ func runConfigInit(cmd *cobra.Command, args []string) error {
 
 	ctx := cmd.Context()
 	if _, err := os.Stat(configPath); err == nil {
-		force, _ := cmd.Flags().GetBool("force")
+		force, err := cmd.Flags().GetBool("force")
+		if err != nil {
+			return fmt.Errorf("failed to get force flag: %w", err)
+		}
 		if !force {
 			return fmt.Errorf("config file already exists at %s (use --force to overwrite)", configPath)
 		}
@@ -152,7 +154,7 @@ func runConfigInit(cmd *cobra.Command, args []string) error {
 		}
 	}
 
-	cfg := runlogic.ConfigFromContext(cmd)
+	cfg := configFromContext(cmd)
 	if cfg == nil {
 		return fmt.Errorf("config not available in context")
 	}
@@ -176,7 +178,7 @@ func runConfigInit(cmd *cobra.Command, args []string) error {
 }
 
 func runConfigShow(cmd *cobra.Command, args []string) error {
-	cfg := runlogic.ConfigFromContext(cmd)
+	cfg := configFromContext(cmd)
 	if cfg == nil {
 		return fmt.Errorf("config not available in context")
 	}
@@ -253,7 +255,7 @@ func runConfigSet(cmd *cobra.Command, args []string) error {
 }
 
 func runConfigGet(cmd *cobra.Command, args []string) error {
-	cfg := runlogic.ConfigFromContext(cmd)
+	cfg := configFromContext(cmd)
 	if cfg == nil {
 		return fmt.Errorf("config not available in context")
 	}
